requirements.txt
anyio==3.6.2
click==8.1.3
dnspython==2.2.1
fastapi==0.88.0
h11==0.14.0
idna==3.4
kafka-python==2.0.2
pydantic==1.10.3
pymongo==4.3.3
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
